refactor and test server logic
