+++ lib/meshchat.rb
@@ -34,6 +34,7 @@ require 'meshchat/config/settings'
 require 'meshchat/net/request'
 require 'meshchat/net/client'
 require 'meshchat/net/listener/request'
+require 'meshchat/net/listener/message_processor'
 require 'meshchat/net/listener/request_processor'
 require 'meshchat/net/listener/server'
 require 'meshchat/cli'
